requirements.txt
ollama==0.1.6
python-dotenv==1.0.1
click==8.1.7
jinja2==3.1.3
pyyaml==6.0.1 